### 2024-06-26 #6933 @crvernon
:wave:  @timothyas -There are a few things that you can take care of until I get a topic editor assigned to your submission:

- [ ] fix your paper.md file header.  I believe you have an extra tab in front of your ORCID specifications.  Once fixed, we should be able to compile from here.
- [ ] Please bring the word count of your paper to around a max of 1000 words.  You can reference your documentation if needed from the paper.
- [ ] I don't see a paper.bib file with your bibliography present in your repo

Thanks!
