### 2024-06-26 9b28f7a
Fix pypi publish and paper tabs(#73)
