### 2024-06-26 #6933 @timothyas
@crvernon thank you for this!

* I think I fixed the tabbing issue in the paper.md header, it looks like editorialbot was able to compile the pdf.
* I am using a bibliography file located within the docs directory of the repo (at [docs/references.bib](https://github.com/timothyas/xesn/blob/main/docs/references.bib)). Will this work, or does it have to be paper.bib in the main directory of the repo?
* I will work on the wordcount, I didn't realize the word limit was strict.
