### 2024-06-27 #6933 @jromanowska
Hi, @crvernon, thanks for the invitation, but I need to decline - I'm going away for holidays entire July. See you in August! :wave:
