+++ paper.md
@@ -66,12 +66,8 @@ signal processing [@jaeger_harnessing_2004].
 
 # Statement of Need
 
-ESNs are a conceptually simple Recurrent Neural Network architecture.
-As shown in the [Implemented Architectures](#implemented-architectures) section,
-they are defined by a simple set of equations, a single hidden layer, and a
-handful of parameters.
-Because of the architecture's simplicity, many scientists who use ESNs implement
-them from scratch.
+ESNs are a conceptually simple Recurrent Neural Network architecture, and
+as a result, many scientists who use ESNs implement them from scratch.
 While this approach can work well for low dimensional problems, the situation
 quickly becomes more complicated when:
 
@@ -90,6 +86,7 @@ design impact study shown by @platt_systematic_2022.
 At its core, `xesn` uses NumPy [@harris_array_2020] and SciPy [@scipy_2020] to perform array based
 operations on CPUs.
 The package then harnesses the CPU/GPU agnostic code capabilities afforded by CuPy [@cupy_learningsys2017]
+and the [`cupy-xarray`](https://cupy-xarray.readthedocs.io/index.html) package
 to operate on GPUs.
 
 ## Parameter Optimization
@@ -97,17 +94,17 @@ to operate on GPUs.
 Although ESNs do not employ backpropagation to train the internal weights of the
 input matrix, adjacency matrix, or bias vector,
 their behavior and performance is highly sensitive to a set of
-5 scalar parameters
-(see [Macro-Scale Parameters](#macro-scale-parameters)).
+5 scalar parameters.
 Moreover, the interaction of these parameters is often not straightforward, and
 it is therefore advantageous to optimize these parameter values, as is shown by
 @platt_systematic_2022 with an extensive set of experimental results.
 Additionally, @platt_constraining_2023 showed that adding invariant metrics to
 the loss function, such as the Lyapunov exponent spectrum or simply its leading element,
 resulted in networks that generalize better to unseen test data.
-@smith_temporal_2023 showed similar results, highlighting that constraining the Kinetic Energy
-spectrum when developing ESN based emulators of Surface Quasi-Geostrophic
-Turbulence helped to preserve small scale features in the flow.
+@smith_temporal_2023 showed similar results in a higher dimensional system using `xesn`'s
+parallelized (lazy) architecture, highlighting that constraining the Kinetic Energy
+spectrum when developing ESN based emulators of turbulent flows helped to
+preserve small scale features.
 As a somewhat generic and efficient implementation of these metrics,
 `xesn` offers the capability to constrain the
 system's Power Spectral Density during parameter optimization, in addition to a
@@ -131,9 +128,8 @@ with scheduling systems like SLURM on HPC environments or in the cloud.
 ## Scaling to Higher Dimensions
 
 It is typical for ESNs to use a hidden layer that is $\mathcal{O}(10-100)$ times
-larger than the input and target space.
-Forecasting large target spaces (e.g., $>\mathcal{O}(10^6)$ in weather and
-climate modeling) quickly becomes intractable with a single reservoir.
+larger than the input and target space, so forecasting large target spaces
+quickly becomes intractable with a single reservoir.
 To address this, @pathak_model-free_2018 developed a parallelization strategy
[…]
