### 2024-06-27 b7e55e9
small changes and add mods to methods
