+++ docs/methods.rst
@@ -19,7 +19,7 @@ Standard ESN Architecture
 #########################
 
 The basic ESN architecture that is implemented by the :class:`xesn.ESN` class
-is defined as follows: 
+is defined by the following discrete timestepping equations: 
 
 .. math::
    \mathbf{r}(n + 1) = (1 - \alpha) \mathbf{r}(n) +
@@ -95,7 +95,7 @@ These options could be selected by choosing:
        "normalization": "eig",
        "is_sparse": True,
        "connectedness": 10,
-       }
+   }
 
 As another example, in :cite:t:`smith_temporal_2023` the authors used a dense
 input matrix with elements randomly chosen from a uniform distribution from -1
@@ -108,7 +108,7 @@ This could be achieved as follows:
        "factor": 0.5,
        "distribution": "uniform",
        "normalization": "svd",
-       }
+   }
 
 with the ``factor=0.5`` just for the sake of an example, and note that
 ``is_sparse=False`` is the default if the option is not provided.
@@ -199,6 +199,7 @@ Example: SQG Turbulence Dataset
 We describe the parallelization strategy based on the dataset used by
 :cite:t:`smith_temporal_2023`, which was generated by
 `this model, written by Jeff Whitaker <https://github.com/jswhit/sqgturb>`_,
+following :cite:`blumen_uniform_1978,held_surface_1995`
 for Surface Quasi-Geostrophic turbulence.
 For the purposes of this discussion, all that matters is the size of the
 dataset, which is illustrated below, and more details can be found in Section 2
@@ -213,7 +214,8 @@ of :cite:t:`smith_temporal_2023`.
 
 The dataset has 3 spatial dimensions :math:`(x, y, z)`, and evolves in time, so
 that the shape is :math:`(N_x = 64, N_y = 64, N_z = 2, N_{time})`.
-We first subdivide the domain into smaller chunks along the :math:`x` and :math:`y`
+Parallelization is achieved in this case by subdividing
+the domain into smaller chunks along the :math:`x` and :math:`y`
 dimensions, akin to domain decomposition techniques in General Circulation
 Models.
 The subdivisions are defined by specifying a chunk size
@@ -343,6 +345,14 @@ Note that the various macro-scale paramaters
 Therefore the only components that drive unique hidden states on each chunk are
 the different input states :math:`\mathbf{u}_k` and the readout matrices
 :math:`\mathbf{W}_\text{out}^k`.
+Additionally, because the solution to the ridge regression problem is linear,
+the readout matrices are simply:
+
+.. math::
+   \mathbf{W}^k_\text{out} = \mathbf{V}_k\mathbf{R}_k^T
+    \left(\mathbf{R}_k\mathbf{R}_k^T + \beta\mathbf{I}\right)^{-1}
+
+where each readout matrix can be computed independently of one another.
 
 Macro-Scale Parameters
 ######################
