### 2024-06-27 fd24869
Reduce Paper Length (#75)
* trying this...

* down to 1600

* small changes and add mods to methods
