+++ paper.md
@@ -66,12 +66,8 @@ signal processing [@jaeger_harnessing_2004].
 
 # Statement of Need
 
-ESNs are a conceptually simple Recurrent Neural Network architecture.
-As shown in the [Implemented Architectures](#implemented-architectures) section,
-they are defined by a simple set of equations, a single hidden layer, and a
-handful of parameters.
-Because of the architecture's simplicity, many scientists who use ESNs implement
-them from scratch.
+ESNs are a conceptually simple Recurrent Neural Network architecture,
+leading many scientists who use ESNs implement them from scratch.
 While this approach can work well for low dimensional problems, the situation
 quickly becomes more complicated when:
 
@@ -94,34 +90,23 @@ to operate on GPUs.
 
 ## Parameter Optimization
 
-Although ESNs do not employ backpropagation to train the internal weights of the
-input matrix, adjacency matrix, or bias vector,
+Although ESNs do not employ backpropagation to train internal weights,
 their behavior and performance is highly sensitive to a set of
-5 scalar parameters
-(see [Macro-Scale Parameters](#macro-scale-parameters)).
+5 scalar parameters.
 Moreover, the interaction of these parameters is often not straightforward, and
-it is therefore advantageous to optimize these parameter values, as is shown by
-@platt_systematic_2022 with an extensive set of experimental results.
-Additionally, @platt_constraining_2023 showed that adding invariant metrics to
-the loss function, such as the Lyapunov exponent spectrum or simply its leading element,
-resulted in networks that generalize better to unseen test data.
-@smith_temporal_2023 showed similar results, highlighting that constraining the Kinetic Energy
-spectrum when developing ESN based emulators of Surface Quasi-Geostrophic
-Turbulence helped to preserve small scale features in the flow.
-As a somewhat generic and efficient implementation of these metrics,
+it is therefore advantageous to optimize these parameter values
+[@platt_systematic_2022].
+Additionally, @platt_constraining_2023 and @smith_temporal_2023 showed that
+adding invariant metrics to the loss function, like the leading Lyapunov
+exponent or the Kinetic Energy spectrum, improved generalizability.
+As a generic implementation of these metrics,
 `xesn` offers the capability to constrain the
 system's Power Spectral Density during parameter optimization, in addition to a
-more traditional
-Normalized Root Mean Squared Error loss function.
+more traditional mean squared error loss function.
 
 `Xesn` enables parameter optimization by integrating with the Surrogate Modeling
 Toolbox [@bouhlel_scalable_2020], which has a Bayesian Optimization
 implementation.
-The parameter optimization process is in general somewhat complex, requiring the user to
-specify a variety of hyperparameters (e.g., the optimization bounds for each
-parameter and the maximum number of optimization steps to take).
-Additionally, for large problems, the process can have heavy compute
-requirements and therefore necessitate HPC or cloud resources.
 `Xesn` provides a simple interface so that the user can specify all of the
 settings for training, parameter optimization, and testing with a single YAML
 file.
@@ -131,9 +116,8 @@ with scheduling systems like SLURM on HPC environments or in the cloud.
 ## Scaling to Higher Dimensions
 
[…]
