### 2024-06-27 #6933 @timothyas
@crvernon I was able to reduce the paper length dramatically from what it was, to ~1600 words. Please let me know if this is acceptable for JOSS. Thank you!
