### 2024-06-27 #6933 @crvernon
Thanks @timothyas - we will let the assigned topic editor and reviewers provide feedback before I ask for any further cuts.  Thanks for knocking this out quickly!

> @editorialbot generate pdf
>
> @crvernon I was able to reduce the paper length dramatically from what it was, to ~1600 words. Please let me know if this is acceptable for JOSS. Thank you!
