### 2024-06-28 #6933 @sneakers-the-rat
It looks like climate modeling is one of the major intended applications - maybe we can find one reviewer who is a climate modeler and maybe the other is a more general neural networks person? Are there any people the authors would like to recommend as reviewers?

Otherwise I see that @SarthakJariwala , @trontrytel , and @Zeitsperre might be a good fit here - are any of y'all interested in reviewing this package?
