### 2024-06-28 #6933 @timothyas
Thank you @sneakers-the-rat! I can recommend Troy Arcomano (GitHub username: [Arcomano1234](https://github.com/Arcomano1234)) who has extensive experience with Echo State Networks and ML more broadly, applied to weather forecasting. His work is cited in the "Previously existing software" section of our paper.
