### 2024-09-25 #7286 @editorialbot
[REVIEW]: xesn: Echo state networks powered by xarray and dask
**Submitting author:** @timothyas (<a href="http://orcid.org/0000-0003-4463-6126">Timothy Smith</a>)
**Repository:** https://github.com/timothyas/xesn
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.1
