+++ docs/references.bib
@@ -21,7 +21,7 @@
 	volume = {11},
 	issn = {1991-959X},
 	url = {https://gmd.copernicus.org/articles/11/3999/2018/},
-	doi = {https://doi.org/10.5194/gmd-11-3999-2018},
+	doi = {10.5194/gmd-11-3999-2018},
 	abstract = {{\textless}p{\textgreater}{\textless}strong class="journal-contentHeaderColor"{\textgreater}Abstract.{\textless}/strong{\textgreater} Can models that are based on deep learning and trained on atmospheric data compete with weather and climate models that are based on physical principles and the basic equations of motion? This question has been asked often recently due to the boom in deep-learning techniques. The question is valid given the huge amount of data that are available, the computational efficiency of deep-learning techniques and the limitations of today's weather and climate models in particular with respect to resolution and complexity.{\textless}/p{\textgreater}{\textless}p{\textgreater}In this paper, the question will be discussed in the context of global weather forecasts. A toy model for global weather predictions will be presented and used to identify challenges and fundamental design choices for a forecast system based on neural networks.{\textless}/p{\textgreater}},
 	language = {English},
 	number = {10},
@@ -41,7 +41,7 @@
 	issn = {1942-2466},
 	shorttitle = {{WeatherBench}},
 	url = {http://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2020MS002203},
-	doi = {https://doi.org/10.1029/2020MS002203},
+	doi = {10.1029/2020MS002203},
 	abstract = {Data-driven approaches, most prominently deep learning, have become powerful tools for prediction in many domains. A natural question to ask is whether data-driven methods could also be used to predict global weather patterns days in advance. First studies show promise but the lack of a common data set and evaluation metrics make intercomparison between studies difficult. Here we present a benchmark data set for data-driven medium-range weather forecasting (specifically 3–5 days), a topic of high scientific interest for atmospheric and computer scientists alike. We provide data derived from the ERA5 archive that has been processed to facilitate the use in machine learning models. We propose simple and clear evaluation metrics which will enable a direct comparison between different methods. Further, we provide baseline scores from simple linear regression techniques, deep learning models, as well as purely physical forecasting models. The data set is publicly available at https://github.com/pangeo-data/WeatherBench and the companion code is reproducible with tutorials for getting started. We hope that this data set will accelerate research in data-driven weather forecasting.},
 	language = {en},
 	number = {11},
@@ -96,6 +96,7 @@
 @article{chen_predicting_2021,
 	title = {Predicting {Shallow} {Water} {Dynamics} using {Echo}-{State} {Networks} with {Transfer} {Learning}},
 	url = {http://arxiv.org/abs/2112.09182},
+    doi = {10.48550/arXiv.2112.09182},
[…]
