+++ docs/references.bib
@@ -640,7 +640,7 @@ This article is part of the theme issue ‘Machine learning for weather and clim
   title={Statistics for spatial data},
   author={Cressie, N},
   year={1993},
-  publisher={Wiley Location New York, NY}
+  publisher={Wiley Location New York, NY},
   doi={10.1002/9781119115151},
 }
 @article{penland_random_1989,
