+++ docs/references.bib
@@ -1562,7 +1562,7 @@ Section: Journal of the Atmospheric Sciences},
   volume={157},
   pages={180},
   year={2006},
-  publisher={Citeseer}
+  publisher={Citeseer},
   doi={10.1017/CBO9780511617652.008},
 }
 
