### 2024-09-26 1524713
Docs: Fix DOIs (#77)
* fix dois noted by editorialbot

* upload-artifact v1 -> v4

* a comma

* comma again
