### 2024-09-26 #7286 @timothyas
Thanks @sneakers-the-rat! I don't have any preferences regarding issue naming conventions. Also, I just want to let everyone know that I'll be out on leave next week, so I may be slow to respond. Additionally, I'm a federal employee at NOAA, and so if the government does shutdown, then I'm legally forbidden from working. So in such an event, I will have to wait until the government opens back up to reply. All that said, please know that I'm very interested in working with you all on this review.

Finally, I updated the paper draft to address DOI issues in [this PR](https://github.com/timothyas/xesn/pull/77). A lot of the citations in the references.bib file are not necessary for the paper, so many don't matter. But I added the ones I could find that are relevant.
