### 2024-09-26 #7286 @sneakers-the-rat
> I updated the paper draft to address DOI issues in https://github.com/timothyas/xesn/pull/77.

Thanks for this, i was going to say something when i saw the DOI report - I will double check with EiC but i believe only the citations that are used in the paper need to have valid DOIs, it looks like you might be using that bibliography in your docs as well, which is totally fine (and imo a good practice to single source them).

Please take your time. I think ideally we don't like to let reviews drag on too long mostly so that they are finished in a reasonable amount of time for authors and to not keep reviewers on the line indefinitely, but if anyone needs extra time, this is all volunteer labor so i certainly won't be upset - key thing to me is just communication (which i modeled poorly at the beginning, and you are modeling excellently now) <3
