### 2024-10-15 #7286 @Arcomano1234
I have completed my review and recommend publication in its current form. The software is well-documented and fills the gap in the lack of GPU-ready software for reservoir computing (RCs) in the Python ecosystem. I look forward to the exciting new research and applications of RCs.

My one suggestion is to allow for the leaking rate to be represented by a matrix of different leaking rates in the reservoir. Specifically a diagonal matrix with a random distribution of leakages. This methodology seems to help with systems with multiple timescales (e.g. see Patel et. al. 2024 https://arxiv.org/pdf/2405.19518 and Tanaka et. al. 2022 https://journals.aps.org/prresearch/abstract/10.1103/PhysRevResearch.4.L032014). I will make this feature request on the xesn repo, but this request should not hold up the publication of this package.
