### 2024-10-16 #7286 @timothyas
Thanks @Arcomano1234! Good suggestion regarding the leak rate. If I recall correctly, my colleagues tried the "multi scale leak rate" and did not notice any impact on the systems they were working with, but it's worth trying again, and we can continue that discussion with the feature request.

Thanks again
