### 2024-10-29 #7286 @timothyas
Hi @sneakers-the-rat, I am wondering if it would be possible to change the license on this package, and if that would slow anything down with the review process? E.g. would the reviewers need to sign off on it again?

I originally thought that we were supposed to publish software under a license like LGPL at NOAA, but I've heard that things are not so restrictive, so I would prefer to use a more permissive license (Apache 2.0). Let me know what this process might look like... Thanks!
