### 2024-10-29 #7286 @sneakers-the-rat
@timothyas the only requirement that JOSS has re: licensing is that it has to be an OSI approved license. I have never heard of a reviewer objecting to the license that a package uses, and I would make a quick check with the reviewers to see if it changes anything about their evaluation, so if you could let us know what change you'd like to make, then we can check in :). I'll do editorial checks now
