### 2024-10-30 timothyas/xesn#78 @sneakers-the-rat
[joss] Editorial paper proofreading requests :)
Part of: https://github.com/openjournals/joss-reviews/issues/7286

Some very small proofreading changes from reading the paper. Overall it's great, just some suggestions that might improve the strength of the paper. The minor suggestions are up to you, take them or leave them, but I would like some clarification on the medium/general suggestions/questions for the purposes of making the paper as strong as it can be :)

Minor/grammatical suggestions:

- L29: "who use ESNs implement them from scratch" -> "who use ESNs **to** implement them from scratch."
- L35: "streamlined, based on the" -> "streamlined based on the"
- L49: "optimization, in addition" -> "optimization in addition"
- L51: "Bayesian Optimization" -> "Bayesian optimization"
- L66-68: list is a bit awkward, maybe something like "map them across available resources, from a laptop to a distributed HPC cluster." - if i'm reading your purpose there correctly re: emphasizing how dask lets you scale from small compute to big compute?
- L79: unclear to me if this is just showing another similar package, or if this is the one that you build `xesn` on top of as mentioned in L62? If it's the former, the wording is a bit awkward - "finally" in a list of two - could be something like "Another implementation is that of Arcomano et. al (2020, 2022, 2023), available at..." If it's the latter, it would be nice to make that more specific :)
- L86: i think this should probably be "the methodology section of the documentation" - in case that link changes, someone can know what page you're referring to.
- Fig 1 caption and elsewhere: "Walltime" -> "Wall time" in the caption (don't worry about changing the labels in the figure) - am aware this is jargon in profiling, but adding the space comes across as mildly less like jargon :)

Medium questions/suggestions:

- Fig 1: I think what you mean with the dotted lines is that those are the theoretical *minimums* for memory? i.e. what one would expect if one were to just multiply the theoretical size of the dtype x params? Caption is not clear on that if so. I am also not sure what the equation is in the legend or how the `a` and `b` constants were empirically derived - would love to see some more description there.
- Fig 1: Are there also dotted lines on the solid orange/green lines? I can't really see those. If the dotted lines overlap with the solid/dashed lines, it would be nice to say so in the caption: "dotted lines are hidden because the measured memory usage visually overlaps with the theoretical minimums"
- Fig 1: Would it be possible to make the visual distinction between dashed vs. solid lines in the figure legend clearer? It just looks like my PDF is having trouble rendering. Maybe decrease the size of the dashes in the legend, or reposition the open parts of the dash so they fall on either side of the center dot rather than almost underneath it?
- Fig 1: Are the memory measurements average or peak memory usages? peak memory use is usually the limiting factor, as it causes the run to crash, so that might be nice to clarify.
- Fig 1: Are the memory measurements during GPU runs just the system memory usage, the GPU memory usage, or both? would like to have clarification in the legend.
- L102 and section: Is "strong scaling" a specific term of art? I'm not sure what it means here.
- L105-106: what do you mean by "more or less fixed?" Help me understand why it might not be fixed. Is it something like theoretically they are the same complexity of problem, but minor variation in some implementation details causes them to be insignificantly different?
- L110: What does the default `LocalCluster` configuration end up meaning for your resources?

General suggestions:

- It would be nice to have a link from the two experiments to the code in the repo used to conduct them - I see the `scaling` directory, and I think a permalink to that would be great for readers who might not have the repo in front of them/if you choose to change repo structure in the future.
- Seeing the GPU wall time flat makes me think that most of that time is spent compiling the GPU code/booting the GPUs - you might consider an additional set of samples with very high `Nr` to show how large the system can scale on GPU. That would be for the purposes of showing off your work, so not required, just thought i'd let you know that's what i assume when i'm reading it and it may make for a stronger conclusion.
