### 2024-10-30 timothyas/xesn#79 @sneakers-the-rat
[joss] Add missing DOI
Part of: https://github.com/openjournals/joss-reviews/issues/7286

Is https://doi.org/10.1017/cbo9780511617652.004 the same as the Lorenz 1996 paper you are citing, or is that a different work? Just want to check to satisfy our very picky DOI checker, close this if those are two different citations.
