### 2024-10-30 timothyas/xesn#79 @sneakers-the-rat
I think [this is the citation for dask](https://proceedings.scipy.org/articles/Majora-7b98e3ed-013)

```
@InProceedings{ matthew_rocklin-proc-scipy-2015,
  author    = { {M}atthew {R}ocklin },
  title     = { {D}ask: {P}arallel {C}omputation with {B}locked algorithms and {T}ask {S}cheduling },
  booktitle = { {P}roceedings of the 14th {P}ython in {S}cience {C}onference },
  pages     = { 126 - 132 },
  year      = { 2015 },
  editor    = { {K}athryn {H}uff and {J}ames {B}ergstra },
  doi       = { 10.25080/Majora-7b98e3ed-013 }
}

```
