### 2024-10-30 #7286 @sneakers-the-rat
Alright @timothyas opened some final issues, then we should be good. Once you address those, can you ping me back here with a zenodo DOI and the final version you're archiving?
