### 2024-10-30 #7286 @timothyas
Thanks @sneakers-the-rat! I will address those issues soon.

Regarding the license, I would like to change the package license to [Apache 2.0](https://github.com/apache/.github/blob/main/LICENSE), which is an [OSI approved license](https://opensource.org/license/apache-2-0).

@Arcomano1234 @wiljnich, would you two be OK with the license switch? Sorry for the last minute change.
