### 2024-10-30 #7286 @wiljnich
Yes, go ahead. I have no concerns with the license change. Thank you.On Oct 30, 2024, at 11:32 AM, Troy Arcomano ***@***.***> wrote:﻿
Hi Tim,
Yes of course the change to a more permissive license is okay.
Best,
Troy Arcomano
