### 2024-10-30 23a74e4
updated text as recommended by @sneakers-the-rat in #78
