+++ paper.md
@@ -215,8 +215,8 @@ This result serves as a motivation for our parallelized architecture.
 In order to evaluate the performance of the parallelized architecture, we take
 the Lorenz96 system with dimension $N_u=256$ and subdivide the domain into
 $N_g = \{2, 4, 8, 16, 32\}$ groups.
-We then fix the reservoir size so that $N_r*N_g = 16,000$, so that the problem
-size is fixed and the timing results reflect strong scaling, i.e.,
+We then fix the problem size such that $N_r*N_g = 16,000$, so that
+the timing results reflect strong scaling, i.e.,
 as determined by Amdahl's Law [@amdahl_1967].
 The training task and resources used are otherwise the same as for the standard
 ESN results shown in \autoref{fig:eager}.
