### 2024-10-30 d3888a8
update markersize, some text
