### 2024-10-30 timothyas/xesn#78 @timothyas
Thanks for these @sneakers-the-rat!

## Minor suggestions

I agreed with all of the changes you suggested, and made those in the text. These suggestions definitely cleared up a few things, thanks! Just to answer a few questions (even though I think you got the gist of what was going on):

> L66-68: list is a bit awkward, maybe something like "map them across available resources, from a laptop to a distributed HPC cluster." - if i'm reading your purpose there correctly re: emphasizing how dask lets you scale from small compute to big compute?

That's correct. I just added "cloud" to your suggestion in order to indicate that it can work in the cloud too.

> L79: unclear to me if this is just showing another similar package, or if this is the one that you build xesn on top of as mentioned in L62?

Yeah it was just to list another implementation, so I went with your wording suggestion. Thanks!

## Medium questions/suggestions:

* Thanks for all of the suggestions regarding the Fig 1 caption, I'll add some answers here and then copy/paste the new text below (which I'll merge once you agree with the changes).
  * The dotted lines are a bit misrepresented by the text - "theoretical" is not the way I should've described them. They are just there to provide the user some guidance on how much memory they might need. I actually derived the lines from the memory usage measurements, using constants that took into account the problem size (input dimension and hidden dimension). Hopefully the new caption clears that up, but let me know what you think.
  * Yeah, those lines were overlapping visually and it was generally confusing. I changed this figure so that the measured values are just dots (CPU) or diamonds (GPU) and the derived lines are solid lines that lay on top. Hopefully this is a bit more clear.
  * I modified the dash frequency, and also changed the markers so they're different between CPU and GPU. Let me know what you think about this.
  * The memory usage is peak memory, definitely a good call to clarify that!
  * Also yep, GPU memory just indicates peak memory usage on the GPU, since that's the usual bottleneck. Another good clarification

  Here's the new figure and caption

![Screenshot 2024-10-30 at 11 50 22 AM](https://github.com/user-attachments/assets/2f597f8c-d084-4f5e-b3b8-00ecbfb7d5cd)

* Regarding strong scaling, yes this is a type of computational scaling. Formally it refers to scaling that can be determined by Amdahl's law, so I added the reference to that original paper to the text. It's a fairly commonly used term, but let me know if you'd like more description there
* My usage of "more or less" is not really appropriate here, so I removed it - sorry about that. The problem size is most definitely fixed. Re-reading this months later, I can only assume that my "more or less" was referring to the fact that the way dask interacts with the problem can be less intuitive, but that's not really relevant in this context.
* The defaults have been added, again good to know (even I've since forgotten...). Here's the new text:
  ```
  On the CPU resource, a GCP `c2-standard-60` instance, the default
  `dask.distributed.LocalCluster` has 6 workers, each with 5 threads.
  On the GPU resource, a GCP `a2-highgpu-8g` instance, the default
  `dask_cuda.LocalCUDACluster` has 8 workers, each with 1 thread.
  ```

## General suggestions

* Definitely agreed regarding the permalink, I added that to the first paragraph in the "Computational Performance" section.
* Thanks for the suggestion on the larger experiments. Unfortunately I don't have enough GCP allocation right now to afford additional tests, even though you're right, this would show off the code more.
