### 2024-10-30 timothyas/xesn#78 @timothyas
FYI @sneakers-the-rat the changes I made to the text can be seen in #80. The paper draft workflow is there, so a pdf can be found, e.g. [here](https://github.com/timothyas/xesn/actions/runs/11598829304?pr=80)
