### 2024-10-30 timothyas/xesn#79 @timothyas
Regarding the Lorenz, 1996 citation, I saw that suggestion, but I don't think it's correct. I also don't have access to that reference, so I'm not sure. This citation is from a conference and is highly cited, but the way that it's cited varies a lot between publications. The citation I gave is the best one that I can come up with unfortunately (it's the one our group has been using).

As for the dask citation, that's interesting. They used to list the citation I gave on their documentation page as "How to Cite", but I don't see that anymore... I'll add the citation you listed as well just to cover my bases here.
