### 2024-10-30 d5a91fd
add dask scipy ref
