+++ docs/references.bib
@@ -1812,3 +1812,13 @@ numpages = {3},
 location = {Atlantic City, New Jersey},
 series = {AFIPS '67 (Spring)}
 }
+
+@InProceedings{rocklin_scipy_2015,
+  author    = { {M}atthew {R}ocklin },
+  title     = { {D}ask: {P}arallel {C}omputation with {B}locked algorithms and {T}ask {S}cheduling },
+  booktitle = { {P}roceedings of the 14th {P}ython in {S}cience {C}onference },
+  pages     = { 126 - 132 },
+  year      = { 2015 },
+  editor    = { {K}athryn {H}uff and {J}ames {B}ergstra },
+  doi       = { 10.25080/Majora-7b98e3ed-013 }
+}
