### 2024-10-31 timothyas/xesn#79 @sneakers-the-rat
> I don't think it's correct.

Got it! you know better than me.

> I don't see that anymore

odd... well when in doubt you can double-cite: citations are free to make and costly (to the cited authors) to miss. neutral on this change, either/or works for me.

feel free to close this once you merge #80
