### 2024-10-31 timothyas/xesn#78 @timothyas
Hi @sneakers-the-rat thanks for all of the comments and feedback! Here's my reply for the last remaining questions/issues.

* On Figure 1, I really appreciate your input, but I respectfully disagree. I think that having the black and gray lines in the plot gives the user an intuitive understanding for how well they can expect the equations to map to their application. I could make the two plots similar by deriving comparable curves for walltime, but I expect this to be less machine independent than memory.

  I am having a change of heart, and will flip the colors to indicate CPU vs GPU, and the markers/linestyle to indicate system state size. But I do strongly prefer having the gray/black lines in the plot. Here's the latest version.

![Screenshot 2024-10-31 at 11 59 38 AM](https://github.com/user-attachments/assets/e810e84a-d08a-475d-9029-b1bb6c693d0d)

> If this is a term of art that you would expect your readers to know, that's good enough for me.

* I went in and added a more intuitive description, and please note that my previous edit included a citation to Amdahl's law (still there)

  ```
  We then fix the problem size such that $N_r*N_g = 16,000$, so that
  the timing results reflect strong scaling.
  That is, the results show how the code performs with increasing resources on a fixed problem
  size, which in theory correspond to Amdahl's Law [@amdahl_1967].
  ```

> As a curiosity question, have you profiled to see if my hunch was right? that once the code starts running it finishes extremely quickly, and most of that time is spent in cuda compilation? or is there some other reason it's flat?

* I think that you're right, that would be my guess too, but I haven't profiled the code to this level of detail.
