### 2024-11-01 timothyas/xesn#78 @timothyas
Hi @sneakers-the-rat, thank you for all of these comments!

> What about using solid/dashed lines instead of black/gray? They're visually distinct enough that it wouldnt get confused with the solid/dotted on the left. Black/gray is just a bit hard to tell apart.

I appreciate the suggestion, but I will keep the lines solid, since the difference between dashed and solid lines indicates the difference between the two system sizes.

> Also since the gray lines follow the CPU, and the black follows GPU, what do you think about replacing the legend labels for those lines with some readable shorthand for the equation like "calculated CPU memory" (or whatever you think is accurate, just an example) and then putting the equation in the caption?

Thanks for the suggestion! I would prefer to keep it how it is, with the equation clearly stated in the legend. I did however modify the following phrase in the caption to be more specific about which line goes to CPU and which to GPU:

_"The gray and black lines indicate the general trend in memory usage for the CPU and GPU simulations, respectively._

Also, I added subtitles to the legend so that it's clear which line goes with CPUs and which goes with GPUs.

> it may be worth saying in the caption that the dots/triangles overlap on the right plot but there are still two curves there

Thanks, I added a white markeredgeborder to the plot to make this more apparent.

---

Here is the latest figure and caption. I hope we have converged, but let me know if you have additional comments.

![Screenshot 2024-11-01 at 9 37 03 AM](https://github.com/user-attachments/assets/b76de5f5-afed-4ada-bd49-7edd39403a05)
