+++ paper.md
@@ -32,7 +32,7 @@ affiliations:
    index: 4
  - name: Pacific Northwest National Laboratory, Richland, WA, USA
    index: 5
-date: 15 December 2023
+date: 1 November 2024
 bibliography: docs/references.bib
 
 ---
@@ -129,7 +129,7 @@ as the basis for `xesn`.
 level operations with `xarray` [@hoyer_xarray_2017].
 As with `xarray`, users refer to dimensions based on their named axes.
 `Xesn` parallelizes the core array based operations by using `dask`
-[@dask_2016,@rocklin_scipy_2015]
+[@dask_2016; @rocklin_scipy_2015]
 to map them across available resources, from a laptop to a distributed HPC or
 cloud cluster.
 
