### 2024-11-01 71c280c
[joss] Paper Edits (#80)
* updated text as recommended by @sneakers-the-rat in #78

* fixed an awkward wording

* updated plot

* update markersize, some text

* add dask scipy ref

* whoops, some license leftovers

* updated fig 1

* strong scaling comment

* another figure iteration

* Update LICENSE (#81)

* Update LICENSE

* bump version

* date, and fix dask double cite

* n_i -> n_u

* another update
