+++ paper.md
@@ -32,7 +32,7 @@ affiliations:
    index: 4
  - name: Pacific Northwest National Laboratory, Richland, WA, USA
    index: 5
-date: 15 December 2023
+date: 1 November 2024
 bibliography: docs/references.bib
 
 ---
@@ -67,7 +67,7 @@ signal processing [@jaeger_harnessing_2004].
 # Statement of Need
 
 ESNs are a conceptually simple Recurrent Neural Network architecture,
-leading many scientists who use ESNs implement them from scratch.
+leading many scientists who use ESNs to implement them from scratch.
 While this approach can work well for low dimensional problems, the situation
 quickly becomes more complicated when:
 
@@ -78,7 +78,7 @@ quickly becomes more complicated when:
 
 `Xesn` is designed to address all of these points.
 Additionally, while there are some design flexibilities for the ESN
-architectures, the overall interface is streamlined, based on the parameter and
+architectures, the overall interface is streamlined based on the parameter and
 design impact study shown by @platt_systematic_2022.
 
 ## GPU Deployment
@@ -101,11 +101,11 @@ adding invariant metrics to the loss function, like the leading Lyapunov
 exponent or the Kinetic Energy spectrum, improved generalizability.
 As a generic implementation of these metrics,
 `xesn` offers the capability to constrain the
-system's Power Spectral Density during parameter optimization, in addition to a
+system's Power Spectral Density during parameter optimization in addition to a
 more traditional mean squared error loss function.
 
 `Xesn` enables parameter optimization by integrating with the Surrogate Modeling
-Toolbox [@bouhlel_scalable_2020], which has a Bayesian Optimization
+Toolbox [@bouhlel_scalable_2020], which has a Bayesian optimization
 implementation.
 `Xesn` provides a simple interface so that the user can specify all of the
 settings for training, parameter optimization, and testing with a single YAML
@@ -128,10 +128,10 @@ as the basis for `xesn`.
 `Xesn` enables prediction for multi-dimensional systems by integrating its high
 level operations with `xarray` [@hoyer_xarray_2017].
 As with `xarray`, users refer to dimensions based on their named axes.
-`Xesn` parallelizes the core array based operations by using `dask` [@dask_2016]
-to map them across available resources, which can include a multi-threaded
-environment on a laptop or single node, or a distributed computing resource
-such as traditional on-premises HPC or in the cloud.
+`Xesn` parallelizes the core array based operations by using `dask`
+[@dask_2016; @rocklin_scipy_2015]
+to map them across available resources, from a laptop to a distributed HPC or
+cloud cluster.
 
 
 ## Existing Reservoir Computing Software
@@ -145,12 +145,11 @@ underlying the reservoir, and allowing for delayed connections.
 On the other hand, `xesn` is focused specifically on implementing ESN
 architectures that can scale to multi-dimensional forecasting tasks.
 Additionally, while `ReservoirPy` enables hyperparameter grid search capabilities
-via Hyperopt [@hyperopt], `xesn` enables Bayesian Optimization as noted above.
+via Hyperopt [@hyperopt], `xesn` enables Bayesian optimization as noted above.
 
-Finally, we note the code base used by [@arcomano_machine_2020;@arcomano_hybrid_2022;@arcomano_hybrid_2023],
+Another ESN implementation is that of [@arcomano_machine_2020;@arcomano_hybrid_2022;@arcomano_hybrid_2023],
 available at [@arcomano_code].
 The code implements ESNs in Fortran, and focuses on using ESNs for hybrid physics-ML modeling.
-
 
 
 # Computational Performance
@@ -162,29 +161,40 @@ scales with increasing hidden and input dimensions.
[…]
