### 2024-11-01 timothyas/xesn#79 @timothyas
Strong agree: it's easy to cite, but is a shame not to be appropriately cited. I added both dask citations in the text.
