### 2024-11-01 #7286 @timothyas
@sneakers-the-rat here's the last bit of info. Please let me know if I need to do anything else. Thanks!

Final version: 0.2.0
Zenodo DOI: [10.5281/zenodo.14026588](https://doi.org/10.5281/zenodo.14026588)

---
