### 2024-11-01 #7286 @sneakers-the-rat
For EIC -  the author [has confirmed](https://github.com/timothyas/xesn/issues/79#issuecomment-2447949304) DOI listed as missing is not the correct citation, and I confirmed that the skip DOIs are either not used or have no DOI.

Thank you to our reviewers for their time and attention given to the work! and thank you to the authors for the gift of this package. Another lovely collaborative review process in the books as far as i'm concerned - we'll hold off final congratulations until an EIC has a look, but wishing you a pre-emptive `recommend-accept` congratulations <3
