### 2024-11-12 86524bc
More JOSS Paper Edits (#82)
* joss paper typos

* upgrade various actions

* trying a different smt version...

* smt version
