### 2024-11-12 02ef51b
Bump version to 0.2.1, fix SMT version in pyproject.toml
