### 2024-11-12 #7286 @timothyas
Hi @crvernon, thanks for these comments. I've taken care of all of them:

* Paper edits taken care of [here](https://github.com/timothyas/xesn/pull/82)
* Updated Zenodo as you mentioned, the DOI is [10.5281/zenodo.14104298](https://doi.org/10.5281/zenodo.14104298)

One issue is that I had to update the package version to 0.2.1 to ensure that my pypi workflow would update with the right version (in the minor paper edit PR, I found that there is a bug in one of the package's dependencies... so had to fix that up...) I hope that's not an issue. Please let me know if there's anything else that needs to be done. Thanks again!
